ram_naam_jaap/requirements.txt
# Core dependencies
Django==5.0.1
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Sessions
redis==5.0.1
django-redis==5.4.0

# Templates
Jinja2==3.1.2
jinja2-time==0.2.0

# Forms & Validation
django-crispy-forms==2.1

# Authentication
django-allauth==0.60.0

# API
djangorestframework==3.14.0

# Frontend
django-compressor==4.4
django-tailwind==3.6.0
django-htmx==1.16.0

# Utilities
Pillow==10.1.0
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Security
django-cors-headers==4.3.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0 